requirements.txt
flask
cloudscraper
websocket-client
redis
